Rename command and put under different parent
