+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/debug/container/ContainerKeyInfoResponse.java
@@ -15,7 +15,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.apache.hadoop.ozone.debug;
+package org.apache.hadoop.ozone.debug.container;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
